requirements.txt
adafruit_blinka==8.58.1
adafruit_circuitpython_ads1x15==2.4.3
board==1.0
Flask==3.1.1
Requests==2.32.3
RPi.GPIO==0.7.1
dotenv==0.9.9
python-dotenv==1.1.0